Implement GetLogicalDrives for Unix
Implement and unify code for getting drives/mount points
