+++ src/Common/src/System/IO/DriveInfoInternal.Win32.cs
@@ -1,16 +1,25 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
+// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using System.Diagnostics;
+using System.Text;
+
 namespace System.IO
 {
-    public static partial class Directory
+    /// <summary>Contains internal volume helpers that are shared between many projects.</summary>
+    internal static partial class DriveInfoInternal
     {
         public static string[] GetLogicalDrives()
         {
             int drives = Interop.mincore.GetLogicalDrives();
             if (drives == 0)
                 throw Win32Marshal.GetExceptionForLastWin32Error();
+
+            // GetLogicalDrives returns a bitmask starting from 
+            // position 0 "A" indicating whether a drive is present.
+            // Loop over each bit, creating a string for each one
+            // that is set.
 
             uint d = (uint)drives;
             int count = 0;
@@ -19,6 +28,7 @@ namespace System.IO
                 if (((int)d & 1) != 0) count++;
                 d >>= 1;
             }
+
             string[] result = new string[count];
             char[] root = new char[] { 'A', ':', '\\' };
             d = (uint)drives;
