+++ src/Common/src/System/IO/DriveInfoInternal.WinRT.cs
@@ -1,0 +1,19 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Diagnostics;
+using System.Text;
+
+namespace System.IO
+{
+    /// <summary>Contains internal volume helpers that are shared between many projects.</summary>
+    internal static partial class DriveInfoInternal
+    {
+        internal static string[] GetLogicalDrives()
+        {
+            // Getting volume information isn't supported on WinRT
+            throw new PlatformNotSupportedException();
+        }
+    }
+}
