+++ src/System.IO.FileSystem/src/System/IO/FileSystem.cs
@@ -43,5 +43,8 @@ namespace System.IO
         public abstract void SetCurrentDirectory(string fullPath);
         public abstract int MaxPath { get; }
         public abstract int MaxDirectoryPath { get; }
+
+        // Volume
+        public abstract string[] GetLogicalDrives();
     }
 }
